Redo hash / cache implementation for crawling (#232)
As mentioned in #233 , we had to redo the caching implementation to
allow cleaning up files in a way that was more reliable and less strict.

To accomplish this I have added more tests for getting the hash from the
ETag header.

In addition to that I have fixed the tests for the robots.txt to not hit
real endpoints and be mocked.